+++ CHANGELOG.md
@@ -10,6 +10,7 @@ announcements on the StageXL forum or use one of the support links below:
 
 #### Pub version NEXT
   * Added ViewportContainer display object.
+  * Added Stage.console to show render statistics.
   * Added support for GraphicsGradient with WebGL renderer.
   * Added support for GraphicsPattern with WebGL renderer.
   * Added StageRenderMode.AUTO_INVALID for lower CPU/GPU load.